refactor: update magalzim to github project owner namespace
