+++ src/main/java/io/github/matheusmfranco/model/Coin.java
@@ -1,0 +1,5 @@
+package io.github.matheusmfranco.model;
+
+public enum Coin implements Lucky {
+    HEADS, TAILS;
+}
