+++ src/main/java/io/github/matheusmfranco/model/Jokenpo.java
@@ -1,0 +1,5 @@
+package io.github.matheusmfranco.model;
+
+public enum Jokenpo implements Lucky {
+    ROCK, PAPER, SCISSOR
+}
