+++ src/main/java/io/github/matheusmfranco/model/OddsOrEvens.java
@@ -1,0 +1,5 @@
+package io.github.matheusmfranco.model;
+
+public enum OddsOrEvens implements Lucky {
+    ODD, EVEN;
+}
