+++ src/main/java/io/github/matheusmfranco/model/ShortResponse.java
@@ -1,0 +1,5 @@
+package io.github.matheusmfranco.model;
+
+public enum ShortResponse implements Lucky {
+    YES, NO;
+}
