+++ src/main/java/io/github/matheusmfranco/play/Alphabet.java
@@ -1,0 +1,10 @@
+package io.github.matheusmfranco.play;
+
+public class Alphabet extends Play {
+
+    public String play() {
+        final int SIZE = 26;
+        return String.valueOf((char) ('A' + randomize(SIZE)));
+    }
+
+}
