+++ src/main/java/io/github/matheusmfranco/play/Dice.java
@@ -1,0 +1,18 @@
+package io.github.matheusmfranco.play;
+
+public class Dice extends Play {
+
+    public String play(Integer sides) {
+        if (sides == null || sides <= 0) {
+            return play();
+        }
+        return Integer.toString(randomize(sides) + 1);
+    }
+
+    @Override
+    public String play() {
+        final int BINARY = 2;
+        return randomize(BINARY).toString();
+    }
+
+}
