+++ src/main/java/io/github/matheusmfranco/play/FlipCoin.java
@@ -1,0 +1,17 @@
+package io.github.matheusmfranco.play;
+
+import io.github.matheusmfranco.model.Coin;
+
+import java.util.Locale;
+
+public class FlipCoin extends FacePlay {
+
+    public FlipCoin(Locale language) {
+        super(language);
+    }
+
+    public String play() {
+        return toSides(Coin.class);
+    }
+
+}
