+++ src/main/java/io/github/matheusmfranco/play/OneTwoGo.java
@@ -1,0 +1,17 @@
+package io.github.matheusmfranco.play;
+
+import io.github.matheusmfranco.model.OddsOrEvens;
+
+import java.util.Locale;
+
+public class OneTwoGo extends FacePlay {
+
+    public OneTwoGo(Locale language) {
+        super(language);
+    }
+
+    public String play() {
+        return toSides(OddsOrEvens.class);
+    }
+
+}
