+++ src/main/java/io/github/matheusmfranco/play/ShakeHand.java
@@ -1,0 +1,17 @@
+package io.github.matheusmfranco.play;
+
+import io.github.matheusmfranco.model.Jokenpo;
+
+import java.util.Locale;
+
+public class ShakeHand extends FacePlay {
+
+    public ShakeHand(Locale language) {
+        super(language);
+    }
+
+    public String play() {
+        return toSides(Jokenpo.class);
+    }
+
+}
