--- src/main/java/org/magalzim/model/Coin.java
@@ -1,5 +1,0 @@
-package org.magalzim.model;
-
-public enum Coin implements Lucky {
-    HEADS, TAILS;
-}
