--- src/main/java/org/magalzim/model/Jokenpo.java
@@ -1,5 +1,0 @@
-package org.magalzim.model;
-
-public enum Jokenpo implements Lucky {
-    ROCK, PAPER, SCISSOR
-}
