--- src/main/java/org/magalzim/model/OddsOrEvens.java
@@ -1,5 +1,0 @@
-package org.magalzim.model;
-
-public enum OddsOrEvens implements Lucky {
-    ODD, EVEN;
-}
