--- src/main/java/org/magalzim/model/ShortResponse.java
@@ -1,5 +1,0 @@
-package org.magalzim.model;
-
-public enum ShortResponse implements Lucky {
-    YES, NO;
-}
