--- src/main/java/org/magalzim/play/Alphabet.java
@@ -1,10 +1,0 @@
-package org.magalzim.play;
-
-public class Alphabet extends Play {
-
-    public String play() {
-        final int SIZE = 26;
-        return String.valueOf((char) ('A' + randomize(SIZE)));
-    }
-
-}
