--- src/main/java/org/magalzim/play/Custom.java
@@ -1,19 +1,0 @@
-package org.magalzim.play;
-
-import java.util.List;
-
-public class Custom extends Play {
-
-    private final List<String> list;
-
-    public Custom(List<String> list) {
-        this.list = list;
-    }
-
-    public String play() {
-        if (list.isEmpty()) {
-            throw new IndexOutOfBoundsException("Index 0 out of bounds for length 0");
-        }
-        return list.get(randomize(list.size()));
-    }
-}
