--- src/main/java/org/magalzim/play/FacePlay.java
@@ -1,30 +1,0 @@
-package org.magalzim.play;
-
-import lombok.AllArgsConstructor;
-
-import org.magalzim.model.Lucky;
-
-import java.util.List;
-import java.util.Locale;
-import java.util.stream.Stream;
-
-@AllArgsConstructor
-public abstract class FacePlay extends Play {
-
-    protected Locale language;
-
-    protected String customLuck(List<String> sides) {
-        if (sides.isEmpty()) {
-            throw new IndexOutOfBoundsException();
-        }
-        return sides.get(randomize(sides.size()));
-    }
-
-    protected <E extends Enum<E> & Lucky>String toSides(Class<E> lucky) {
-        List<String> sides = Stream.of(lucky.getEnumConstants())
-                .map(constant -> constant.fetchTranslation(language))
-                .toList();
-        return customLuck(sides);
-    }
-
-}
